[Protocol] Make quake queries very modular
